Split skills.rs and rules.rs tests into sibling modules
Both files sat exactly at their grandfathered file-size ceilings (1502 and
1666) with zero headroom, and Phase 3 (#714) has to add to both: the skills
miner grows a typed proposal path, and the rules miner — dead since the shared
clustering module was extracted — has to be wired to a caller.

The baseline file forbids ADDING an entry for a new file, so growing these
in place would have meant raising two ceilings. Extracting the inline
`mod tests` to a sibling `tests.rs` is the sanctioned alternative: it is a
pure move (verbatim, dedented one level), and it drops both files off the
grandfather list entirely rather than nudging their ceilings up.

No behavior change. 35 skills + 49 rules tests still pass, unmodified.

Refs #714
